Style: PerformanceInfo 스타일 분리
